refactor: move PdfPasswords to pdf module
